ocr-mosip-integration/enhanced-ocr-module/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
pytesseract==0.3.10
google-cloud-vision==3.4.4
boto3==1.34.0
pymongo==4.6.0
psycopg2-binary
pydantic==2.5.0
python-multipart==0.0.6
Pillow==10.1.0
openai==1.3.5
numpy==1.24.3
python-dotenv==1.0.0
pydantic-settings==2.1.0